backend/requirements.txt
fastapi==0.68.1
uvicorn==0.14.0
asyncpg==0.23.0
tortoise-orm==0.17.4
aerich==0.5.3
pytest==6.2.4
requests==2.25.1
gunicorn==20.1.0
pydantic[email,dotenv]==1.8.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-multipart==0.0.5
pytest-asyncio==0.15.1
fastapi-mail==1.0.0